# docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for managing Pointlings, users, items, and XP system",
        "title": "Pointlings API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/points/{userID}/history": {
            "get": {
                "description": "Get paginated history of a user's point spending transactions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Points"
                ],
                "summary": "Get point spending history",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Number of records to return (max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "Number of records to skip",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.PointSpend"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid user ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/points/{userID}/spend": {
            "post": {
                "description": "Purchase an item for a pointling using user's points",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Points"
                ],
                "summary": "Spend points on an item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Points spending request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.spendPointsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TransactionSuccess"
                        }
                    },
                    "400": {
                        "description": "Invalid user ID or request body",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "402": {
                        "description": "Insufficient points",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Item not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Item already owned",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "handlers.spendPointsRequest": {
            "type": "object",
            "properties": {
                "item_id": {
                    "type": "integer",
                    "example": 123
                },
                "pointling_id": {
                    "type": "integer",
                    "example": 456
                }
            }
        },
...
        "models.Item": {
            "type": "object",
            "properties": {
                "asset_id": {
                    "type": "string"
                },
                "category": {
                    "$ref": "#/definitions/models.ItemCategory"
                },
                "item_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price_points": {
                    "type": "integer"
                },
                "rarity": {
                    "$ref": "#/definitions/models.ItemRarity"
                },
                "slot": {
                    "$ref": "#/definitions/models.ItemSlot"
                },
                "unlock_level": {
                    "type": "integer"
                }
            }
        },
        "models.ItemCategory": {
            "type": "string",
            "enum": [
                "ACCESSORY",
                "FEATURE"
            ],
            "x-enum-varnames": [
                "CategoryAccessory",
                "CategoryFeature"
            ]
        },
        "models.ItemRarity": {
            "type": "string",
            "enum": [
                "COMMON",
                "RARE",
                "EPIC",
                "LEGENDARY"
            ],
            "x-enum-varnames": [
                "RarityCommon",
                "RarityRare",
                "RarityEpic",
                "RarityLegendary"
            ]
        },
        "models.ItemSlot": {
            "type": "string",
            "enum": [
                "HAT",
                "SHOES",
                "FACE",
                "WINGS"
            ],
            "x-enum-varnames": [
                "SlotHat",
                "SlotShoes",
                "SlotFace",
                "SlotWings"
            ]
        },
...
        "models.PointSpend": {
            "type": "object",
            "properties": {
                "item": {
                    "description": "Joined data",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Item"
                        }
                    ]
                },
                "item_id": {
                    "type": "integer"
                },
                "points_spent": {
                    "type": "integer"
                },
                "spend_id": {
                    "type": "integer"
                },
                "spend_ts": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.TransactionSuccess": {
            "type": "object",
            "properties": {
                "item_id": {
                    "type": "integer"
                },
                "new_balance": {
                    "type": "integer"
                },
                "points_spent": {
                    "type": "integer"
                },
                "previous_spend": {
                    "type": "integer"
                }
            }
        },
...
    },
...
}